Add mission history builder

+++ ODMissionStacker/Missions/StackInfo.cs
@@ -6,14 +6,14 @@ namespace ODMissionStacker.Missions
     {
         private string issuingFaction;
         private string targetFaction;
-        private int reward;
+        private long reward;
         private int killCount;
         private int difference;
         private int missionCount;
 
         public string IssuingFaction { get => issuingFaction; set { issuingFaction = value; OnPropertyChanged(); } }
         public string TargetFaction { get => targetFaction; set { targetFaction = value; OnPropertyChanged(); } }
-        public int Reward { get => reward; set { reward = value; OnPropertyChanged(); } }
+        public long Reward { get => reward; set { reward = value; OnPropertyChanged(); } }
         public int KillCount { get => killCount; set { killCount = value; OnPropertyChanged(); } }
         public int Difference { get => difference; set { difference = value; OnPropertyChanged(); } }
         public int MissionCount { get => missionCount; set { missionCount = value; OnPropertyChanged(); } }
